remove redundant api_urls.py
